11.07.21

+++ asp.net/AlApar/AlApar/Models/Animal/AnimalPhoto.cs
@@ -8,7 +8,7 @@ namespace AlApar.Models.Animal
     {
         public int Id { get; set; }
         public string ImagePath { get; set; }
-        public int? AdId { get; set; }
+        public long? AdId { get; set; }
         public int? PrimaryImage { get; set; }
         public string Thumbnail { get; set; }
