+++ asp.net/AlApar/AlApar/Models/Job/JobEmployeePhoto.cs
@@ -10,7 +10,7 @@ namespace AlApar.Models.Job
     {
         public int Id { get; set; }
         public string ImagePath { get; set; }
-        public int? AdId { get; set; }
+        public long? AdId { get; set; }
         public int? PrimaryImage { get; set; }
         public string Thumbnail { get; set; }
